Add files via upload
The solution here is to have an agent bot (or an interface bot) and a few other bots which can handle conversations for a specific domain - let's call these specific bots. The agent bot knows about the specific bots and also about which domain each of them can handle. When user initiates conversation with agent bot, the agent bot will understand the intent of user query and it will redirect the user query to a specific bot. Subsequent requests from user are redirected to specific bot. When the conversation with the specific bot is over or when the specific bot is not able to handle the request, the control is given back to agent bot which will then redirect the messages to appropriate bot.

This approach provides seamless experience for users. It can be used by organisations which provide a host of services to its customers like financial services, tours and travel agencies, news agencies etc..

Advantages with this approach are:

Plug and play the bots
Modular approach facilitates Bots composition
Come up with new services by composing two or more Bots
Easy to maintain, make changes, add/remove functionalities
Easy to troubleshoot issues
Transparent to user
In this code pattern we will use Watson assistant Bot for building Bots and Node.js application as orchestration layer.